requirements.txt
Flask>=2.0.1
Pillow>=9.0.0
rembg>=2.0.55
onnxruntime>=1.16.0
